reorganized models package so that each major class is in a separate module
