+++ mbrl/models/__init__.py
@@ -1,13 +1,18 @@
-from .base_models import BasicEnsemble, Model
-from .dynamics_models import DynamicsModelTrainer, DynamicsModelWrapper
+from .basic_ensemble import BasicEnsemble
+from .common import EnsembleLinearLayer, truncated_normal_init
+from .dynamics_models import DynamicsModelWrapper
 from .gaussian_mlp import GaussianMLP
+from .model import Model
 from .model_env import ModelEnv
+from .model_trainer import DynamicsModelTrainer
 
 __all__ = [
     "Model",
     "BasicEnsemble",
     "DynamicsModelTrainer",
     "DynamicsModelWrapper",
+    "EnsembleLinearLayer",
     "ModelEnv",
     "GaussianMLP",
+    "truncated_normal_init",
 ]
