+++ mbrl/models/basic_ensemble.py
@@ -1,253 +1,12 @@
-import abc
 import warnings
-from typing import List, Optional, Sequence, Tuple, Union, cast
+from typing import Optional, Sequence, Tuple, Union
 
-import hydra.utils
-import numpy as np
+import hydra
 import omegaconf
 import torch
-from torch import nn as nn
+import torch.nn as nn
 
-import mbrl.logger
-import mbrl.math
-import mbrl.types
-
-
-def truncated_normal_init(m: nn.Module):
-    """Initializes the weights of the given module using a truncated normal distribution."""
-
-    if isinstance(m, nn.Linear):
-        input_dim = m.weight.data.shape[0]
-        stddev = 1 / (2 * np.sqrt(input_dim))
-        mbrl.math.truncated_normal_(m.weight.data, std=stddev)
-        m.bias.data.fill_(0.0)
-    if isinstance(m, EnsembleLinearLayer):
-        num_members, input_dim, _ = m.weight.data.shape
-        stddev = 1 / (2 * np.sqrt(input_dim))
-        for i in range(num_members):
-            mbrl.math.truncated_normal_(m.weight.data[i], std=stddev)
-        m.bias.data.fill_(0.0)
-
-
-class EnsembleLinearLayer(nn.Module):
-    """Efficient linear layer for ensemble models."""
-
-    def __init__(
-        self, num_members: int, in_size: int, out_size: int, bias: bool = True
-    ):
-        super().__init__()
-        self.num_members = num_members
-        self.in_size = in_size
-        self.out_size = out_size
-        self.weight = nn.Parameter(
-            torch.rand(self.num_members, self.in_size, self.out_size)
-        )
-        if bias:
-            self.bias = nn.Parameter(torch.rand(self.num_members, 1, self.out_size))
-            self.use_bias = True
-        else:
-            self.use_bias = False
-
-        self.elite_models: List[int] = None
-        self.use_only_elite = False
-
-    def forward(self, x):
-        if self.use_only_elite:
-            xw = x.matmul(self.weight[self.elite_models, ...])
-            if self.use_bias:
-                return xw + self.bias[self.elite_models, ...]
-            else:
-                return xw
-        else:
-            xw = x.matmul(self.weight)
-            if self.use_bias:
-                return xw + self.bias
-            else:
-                return xw
-
-    def extra_repr(self) -> str:
-        return (
-            f"num_members={self.num_members}, in_size={self.in_size}, "
-            f"out_size={self.out_size}, bias={self.use_bias}"
-        )
-
-    def set_elite(self, elite_models: Sequence[int]):
-        self.elite_models = list(elite_models)
-
-    def toggle_use_only_elite(self):
-        self.use_only_elite = not self.use_only_elite
-
-
-# ------------------------------------------------------------------------ #
-# Abstract model
-# ------------------------------------------------------------------------ #
-class Model(nn.Module, abc.ABC):
-    """Base abstract class for all dynamics models.
-
-    All classes derived from `Model` must implement the following methods:
-
-        - ``forward``: computes the model output.
-        - ``loss``: computes a loss tensor that can be used for backpropagation.
-        - ``eval_score``: computes a non-reduced tensor that gives an evaluation score
-          for the model on the input data (e.g., squared error per element).
-        - ``save``: saves the model to a given path.
-        - ``load``: loads the model from a given path.
-
-    Args:
-        in_size (int): size of the input tensor.
-        out_size (int): size of the output tensor.
-        device (str or torch.device): device to use for the model.
-    """
-
-    def __init__(
-        self,
-        in_size: int,
-        out_size: int,
-        device: Union[str, torch.device],
-        *args,
-        **kwargs,
-    ):
-        super().__init__()
-        self.in_size = in_size
-        self.out_size = out_size
-        self.device = torch.device(device)
-        self.to(device)
-
-    def forward(self, x: torch.Tensor, **kwargs) -> Tuple[torch.Tensor, torch.Tensor]:
-        """Computes the output of the dynamics model.
-
-        Args:
-            x (tensor): the input to the model.
-
-        Returns:
-            (tuple of two tensor): the predicted mean and  log variance of the output.
-            If the model does not predict uncertainty, the second output must be ``None``.
-        """
-        pass
-
-    @abc.abstractmethod
-    def loss(
-        self,
-        model_in: torch.Tensor,
-        target: torch.Tensor,
-    ) -> torch.Tensor:
-        """Computes a loss that can be used to update the model using backpropagation.
-
-        Args:
-            model_in (tensor): the inputs to the model.
-            target (tensor): the expected output for the given inputs.
-
-        Returns:
-            (tensor): a loss tensor.
-        """
-
-    @abc.abstractmethod
-    def eval_score(self, model_in: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
-        """Computes an evaluation score for the model over the given input/target.
-
-        This method should compute a non-reduced score for the model, intended mostly for
-        logging/debugging purposes (so, it should not keep gradient information).
-        For example, the following could be a valid
-        implementation of ``eval_score``:
-
-        .. code-block:: python
-
-           with torch.no_grad():
-               return torch.functional.mse_loss(model(model_in), target, reduction="none")
-
-
-        Args:
-            model_in (tensor or sequence of tensors): the inputs to the model
-                                                      (or ensemble of models).
-            target (tensor or sequence of tensors): the expected output for the given inputs.
-
-        Returns:
-            (tensor): a non-reduced tensor score.
-        """
-
-    @abc.abstractmethod
-    def save(self, path: str):
-        """Saves the model to the given path. """
-
-    @abc.abstractmethod
-    def load(self, path: str):
-        """Loads the model from the given path."""
-
-    @abc.abstractmethod
-    def _is_deterministic_impl(self):
-        # Subclasses must specify if model is _deterministic or not
-        pass
-
-    @abc.abstractmethod
-    def _is_ensemble_impl(self):
-        # Subclasses must specify if they are ensembles or not
-        pass
-
-    @property
-    def is_deterministic(self):
-        """Whether the model is deterministic or not."""
-        return self._is_deterministic_impl()
-
-    @property
-    def is_ensemble(self):
-        """Whether the model is an ensemble or not."""
-        return self._is_ensemble_impl()
-
-    def update(
-        self,
-        model_in: torch.Tensor,
-        target: torch.Tensor,
-        optimizer: Union[torch.optim.Optimizer, Sequence[torch.optim.Optimizer]],
-    ) -> float:
-        """Updates the model using backpropagation with given input and target tensors.
-
-        Provides a basic update function, following the steps below:
-
-        .. code-block:: python
-
-           optimizer.zero_grad()
-           loss = self.loss(model_in, target)
-           loss.backward()
-           optimizer.step()
-
-
-        Returns the numeric value of the computed loss.
-
-        """
-        assert not isinstance(optimizer, Sequence)
-        optimizer = cast(torch.optim.Optimizer, optimizer)
-        self.train()
-        optimizer.zero_grad()
-        loss = self.loss(model_in, target)
-        loss.backward()
-        optimizer.step(None)
-        return loss.item()
-
-    def __len__(self):
-        return None
-
-    def sample_propagation_indices(
-        self, batch_size: int, rng: torch.Generator
-    ) -> Optional[torch.Tensor]:
-        """Samples propagation indices used for "fixed_model" propagation.
-
-        This method should be overridden by all ensemble classes, so that indices for
-        "fixed_model" style propagation are sampled (equivalent to TSinf propagation in the
-        PETS paper). This allow each type of ensemble to have its own propagation logic.
-
-        Args:
-            batch_size (int): the batch size to use for the indices.
-            rng (torch.Generator): random number generator.
-        """
-        if self.is_ensemble:
-            raise NotImplementedError(
-                "This method must be implemented by all ensemble classes."
-            )
-        return None
-
-    def set_elite(self, elite_models: Sequence[int]):
-        """For ensemble models, indicates if some models should be considered elite."""
-        pass
+from .model import Model
 
 
 class BasicEnsemble(Model):
