+++ mbrl/models/common.py
@@ -1,0 +1,74 @@
+from typing import List, Sequence
+
+import numpy as np
+import torch
+from torch import nn as nn
+
+import mbrl.logger
+import mbrl.math
+import mbrl.types
+
+
+def truncated_normal_init(m: nn.Module):
+    """Initializes the weights of the given module using a truncated normal distribution."""
+
+    if isinstance(m, nn.Linear):
+        input_dim = m.weight.data.shape[0]
+        stddev = 1 / (2 * np.sqrt(input_dim))
+        mbrl.math.truncated_normal_(m.weight.data, std=stddev)
+        m.bias.data.fill_(0.0)
+    if isinstance(m, EnsembleLinearLayer):
+        num_members, input_dim, _ = m.weight.data.shape
+        stddev = 1 / (2 * np.sqrt(input_dim))
+        for i in range(num_members):
+            mbrl.math.truncated_normal_(m.weight.data[i], std=stddev)
+        m.bias.data.fill_(0.0)
+
+
+class EnsembleLinearLayer(nn.Module):
+    """Efficient linear layer for ensemble models."""
+
+    def __init__(
+        self, num_members: int, in_size: int, out_size: int, bias: bool = True
+    ):
+        super().__init__()
+        self.num_members = num_members
+        self.in_size = in_size
+        self.out_size = out_size
+        self.weight = nn.Parameter(
+            torch.rand(self.num_members, self.in_size, self.out_size)
+        )
+        if bias:
+            self.bias = nn.Parameter(torch.rand(self.num_members, 1, self.out_size))
+            self.use_bias = True
+        else:
+            self.use_bias = False
+
+        self.elite_models: List[int] = None
+        self.use_only_elite = False
+
+    def forward(self, x):
+        if self.use_only_elite:
+            xw = x.matmul(self.weight[self.elite_models, ...])
+            if self.use_bias:
+                return xw + self.bias[self.elite_models, ...]
+            else:
+                return xw
+        else:
+            xw = x.matmul(self.weight)
+            if self.use_bias:
+                return xw + self.bias
+            else:
+                return xw
+
+    def extra_repr(self) -> str:
+        return (
+            f"num_members={self.num_members}, in_size={self.in_size}, "
+            f"out_size={self.out_size}, bias={self.use_bias}"
+        )
+
+    def set_elite(self, elite_models: Sequence[int]):
+        self.elite_models = list(elite_models)
+
+    def toggle_use_only_elite(self):
+        self.use_only_elite = not self.use_only_elite
